+++ helix-contract/contracts/mapping-token/test/MockOutboundLane.sol
@@ -27,7 +27,7 @@ contract MockOutboundLane is MockMessageVerifier {
         remoteInboundLane = _remoteInboundLane;
     }
 
-    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint256) {
+    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint64) {
         // call target contract
         bool result = MockInboundLane(remoteInboundLane).mock_dispatch(msg.sender, targetContract, encoded);
         outboundLaneNonce.latest_generated_nonce += 1;
